added static_cast

+++ dev/get_prepared_statement.h
@@ -125,7 +125,7 @@ namespace sqlite_orm {
         using expression_type = typename statement_type::expression_type;
         using node_tuple = typename internal::node_tuple<expression_type>::type;
         using bind_tuple = typename internal::bindable_filter<node_tuple>::type;
-        using result_tupe = typename std::tuple_element<N, bind_tuple>::type;
+        using result_tupe = typename std::tuple_element<static_cast<size_t>(N), bind_tuple>::type;
         const result_tupe* result = nullptr;
         auto index = -1;
         internal::iterate_ast(statement.t, [&result, &index](auto& node) {
@@ -148,7 +148,7 @@ namespace sqlite_orm {
         using expression_type = typename statement_type::expression_type;
         using node_tuple = typename internal::node_tuple<expression_type>::type;
         using bind_tuple = typename internal::bindable_filter<node_tuple>::type;
-        using result_tupe = typename std::tuple_element<N, bind_tuple>::type;
+        using result_tupe = typename std::tuple_element<static_cast<size_t>(N), bind_tuple>::type;
         result_tupe* result = nullptr;
         auto index = -1;
         internal::iterate_ast(statement.t, [&result, &index](auto& node) {
